Added (almost) all structures, needs and some ships.
Smaller bugfixes.
Colored ComboBox for some inputs.

+++ src/com/mortmann/andja/creator/unitthings/Ship.java
@@ -11,7 +11,7 @@ import com.mortmann.andja.creator.util.Tabable;
 @Root(strict=false)
 public class Ship extends Unit {
 	@FieldInfo(required=true, IsEffectable=true) @Element public int maximumAmountOfCannons = 0;
-	@FieldInfo(required=true, IsEffectable=true) @Element public int damagePerCannon = 1;
+	@FieldInfo(required=true, IsEffectable=true) @Element public float damagePerCannon = 1;
 	@FieldInfo(required=true) @Element public Item cannonType; // WARNING: this has to be filled when maximumAmountOfCannons > 0
 	
 	@Override
